Version Final

+++ src/main/java/com/proyecto/interfaces/CarroCarrera.java
@@ -3,5 +3,5 @@ package com.proyecto.interfaces;
 public interface CarroCarrera {
 
 
-    public int acelerarCarro();
+    public double acelerarCarro();
 }
